Set values of most inputs in Shipping, PMT to state values
State values change based on what is input. This is done for all fields on Payment page, except for expiry date & CVV, as, in the real world, it wouldn't be prudent to keep ALL card info displayed on a page. All this keeps fields filled out if user goes back & forth b/t pgs, so they don't have to refill fields.